Create index_storage directoty if missing

+++ setup_autokpkg_chatbot.sh
@@ -4,5 +4,9 @@ SOURCE=${BASH_SOURCE[0]}
 THISDIR=$(dirname "${SOURCE}")
 cd "$THISDIR"
 rm -r documents/*
-rm -r index_storage/*
+if [ ! -d index_storage ] ; then
+    mkdir index_storage
+else
+    rm -r index_storage/*
+fi
 cp -R example_docs/autopkg/* documents/
